requirements.txt
streamlit
pandas
matplotlib
statsmodels
scikit-learn
numpy
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
